pyproject.toml
[tool.poetry]
name = "kwiz-kreator"
version = "0.1.0"
description = "An editor for creating Trail Trivia quizzes"
authors = ["Inu Labs <91296986+john-need@users.noreply.github.com>"]
license = "Apache 2.0"
readme = "README.md"
packages = [
    { include = "main.py" },
]


[tool.poetry.dependencies]
python = "^3.11"
pyqt5-sip = "^12.13.0"
toolz = "^0.12.0"
pyqt5-stubs = "^5.15.6.0"
pyqt5-qt5 = "5.15.2"
pyqt5 = "^5.15.10"
pyspellchecker = "^0.7.2"
flask = "^3.0.0"
pysftp = "^0.2.9"
language-tool-python = "^2.7.1"
python-debouncer = "^0.1.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
